Initial Release
This app allows you to copy a floppy or hard drive, using the BIOS read services, on one machine and copy this data to another via a standard Serial Null Mode connection.  This is very useful for older computers that do not have a modern external media source such as a USB port.